GhostPDF - yet more circular references
Bug #705724, 705725

The code for checking the Info dictionary for circular references used
pdfi_array_fetch() which carefully ensures the object number of the
returned object doesn't get added to the loop detection array, and
therefore defeats the detection.

I don't want to modify the basic function (because its used a lot) so
I've added a new function which is practically identical but checks the
returned object to see if its a circular reference. The basic problem
is that we can't tell if the returned object was an indirect reference
(and so has been checked already) or was already stored in the array
directly and therefore needs checking. So having the fetch code do the
checking (because it knows which object type it was dealing with) is
a better place to do it.