Fonts - fix cleanup of copied fonts on error
Showed up with some of the OSS-fuzz tests which exhaust memory, when
run with pdfwrite. Example command line:

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dSAFER -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sOutputFile=/temp/out.pdf -sDEVICE=pdfwrite D:\bugs\oss-fuzz\49803.pdf

The problem is that after we allocate the font structure for the copy
of the font, it will be finalized if we free it, so we need to set up
enough of the font immediately, in case we do free it, as noted in the
comment. But we were copying the UID from the source font and not
replacing it until much later.

If we were able to allocate the 'copied' font, but not able to allocate
one of the other structures (glyphs or cfdata) we would goto the fail
condition where we would try to free the 'copied' font. Because we had
not updated the UID this would then crash.

Fix it by moving the UID setup into the section where we initialise
'just enough' of the new font. If setting up the UID fails, then set
the UID to 'invalid' instead.