GhostPDF - Improve handling of empty objects
Bug #705734 " stack overflow in psi/zpdfops.c:79 PDFdict_to_PSdict"

The problem here is the file has an object definition with no endobj
and no actual content:

4 0 obj
5 0 obj

The code to deal with missing endobj tokenswas assuming that if there
were any objects on the stack that these represented the content of the
object.

In this case, that isn't true, though there *are* objects on the stack
they didn't come from the object. For various reasons it just so happens
that this ends up defining the content of object 4 as an indirect
reference to object 4 which is ordinarily impossible, only arrays and
dictionaries can contain indirect references, not objects.

This caused a circular reference that shouldn't be possible.

Check the stack depth on entry and use the current stack depth minus
the initial value instead of simply using the current stack depth. This
then correctly returns a stack underflow error and the circular
reference doesn't happen.