GhostPDF - Don't follow Parent key in Pages dictionary for Resources
When finding a named resource we check the current dictionary (which
may be a Pattern, Xobject, Page etc) and if we fail to find the resource
we look to see if the dictionary has a Parent entry (for Resource
inheritance) and look in that.

If the current dictionary is of /Type /Page, don't do that, because the
Parent will be the Pages tree and we can end up creating a cyclic
reference between the Pages tree and the page dictionary.

Showed up as a memory leak (because we failed to free the page
dictionary) with a few OSS-fuzz files.