Shadings - initialise a struct member in case of later error
OSS-fuzz #50113

If alloc_patch_fill_memory() fails, we would return to the caller, which
would then free the patch_fill_state, which would try to dereference
the 'pcic' member in order to remove a colour from the ICC cache.
But the colour had not been added to the cache and pcic was
uninitialised, resulting in a seg fault.