pdfwrite - improve error handling
OSS-fuzz #50113 - continued

After fixing the initial problem with this file, more crashes emerged
due to pdfwrite (and to some extent the graphics library) poor error
handling.

We were getting to pdf_ferror() having already closed pdev->streams,strm
which had closed the underlying 'file' member and set it to NULL.

sflush(), rather horrifyingly but typical of the stream code, does not
check to see if the stream 'file' member is valid, it simply uses it
as an argument to gp_fflush(). Passing a NULL pointer causes a crash.

Fix it here by checking the file member is not NULL before flushing
the stream (there really is no point flushing it if the file member is
NULL).