pdfwrite - better error handling with CharProcs
Oss-fuzz #50113 - continued

The pdf_copy_mono() function could call pdf_begin_char_proc() which
opens an 'aside' (messes with the stream pointers). But if a later
error occurred it did not close the aside, which caused the streams to
be in the wrong state, and eventually closed the asides stream when it
should not have been leading to a crash.

If we successfully call pdf_begin_char_proc() we now note that in a
boolean, and all return paths after that go through a check to see
if we need to call pdf_end_char_proc() in order to reset the streams.