Check for "make" before checking for GNU make
If left to its defaults, the build system on Unix will assume GNU make, and use
GNU make features to speed up the build process.

configure would check for GNU make for this reason, but wasn't checking if there
was a "make" program available first, leading to a potentiall confusing error
message.

This checks that a "make" program exists, before trying to check it's GNU make.

Secondly, in recognition that some systems (Solaris, for example) have their own
make, but often have GNU make available as "gmake", add a "--with-make=" option
allowing a user to set a custom name for the make executable to look for and
check the version.