GhostPDF - remove Parent key from stream dictionaries after insertion
Bug #705746 "stack-overflow in pdf_misc.c:162 pdfi_string_from_name"

When handling XObjects we check the stream dictionary to see if it is
the same as the one we were handed, or different (Forms). If it is
different then we add the current stream dictionary to the XObject
dictionary as a /Parent, so that Resource searching will trickle back
up the stack of objects until it either finds the Resource or runs out
of dictionaries to check.

However, we didn't remove the Parent again; this turns out to be a
problem if the file is broken in such a way as to define the /Info
dictionary as being one of these streams. If we try to turn the Info
dictionary into a PostScript dictionary after we've inserted a /Parent
then we can end up with a circular reference.

The obvious solution is to remove the key if we added it, we only need
it for the duration of the Form in order to find Resources.