graphics library - curve flattening failure
OSS-fuzz #50243

The fuzzed file strokes a massive curve section, so long that we
decide we have to subdivide it before flattening it. This led to us
failing to satisfy the exit condition, and running an infinite loop
until we ran out of memory trying to add line segments.

Fix the exit test so that we exit if we have no more points to consider.