GhostPDF - Correct Widget annotation key name
The comment around line 3662, copied from the original PostScript
implementation, refers to the FT (Field Type) key, but the code was
actually trying to find a key named /TF, whcih didn't exist because
there is no such key defined.

Fix the spelling, and change all the variable names to match.