GhostPDF - more work on Widget annotations
The first problem is that the code did not do what the comments said.
We exited the loop searching for T and FT keys as soon as we found
either, and then only checked if we had at least one of the keys.

The comments (and the old PostScript code) say we should not render the
annotation unless both keys are present.

Change the loop and exit conditions to implement this.

The second change is handling the /Parent. If we use a simple dictionary
accessor then the value of the key will be dereferenced and stored in
the dictionary. Fine if the Parent is valid but if the Parent value is
an indirect reference to (eg) the Pages tree, then we will end up
with a circular reference (Pages->Page->annot->Pages) which will leak
memory.

Use more complicated code to access the dictionary without storing the
dereferenced value of /Parent in the dictionary.