graphics library - NULL return pointer if ToUnicode allocation fails
Seen with the file from OSS-fuzz 42916 and :
-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dSAFER -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sDEVICE=pdfwrite -sOutputFile=out.pdf

If we failed to allocate 'map' we would properly free the (already
allocated) pcmap, but we left the pointer pointing at the freed memory.
Later we would free the font, and would then attempt to free the free
memory, and crash.

Fix by setting the returned pointer to NULL.