pdfwrite - check allocated memory
Observed with OSS-fuzz 50113 and this command line:

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dSAFER -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sDEVICE=pdfwrite -sOutputFile=out.pdf

We tried to allocate memory to hold a number of Unicode code points, but
we didn't check to see if the allocation succeeded. In this case it
fails, and we then tried to use the pointer.

Check the allocation and return an error if it fails.