GhostPDF - fix memory leak with Widget annotations
Exhibited by /tests_private/pdf/sumatra/1348_-_support_Additional_Actions.pdf

The problem can occur if we find either /T or /FT in the current
dictionary but don't find both. In that case we try to find the keys in
any /Parent dictionary. But if the Parent contains *both* keys then it
would replace the stored one found in the earlier dictionary, without
counting it down, leading to a memory leak.

We don't actually use the /T and /FT values which we retrieve, so the
simplest solution is to simply discard them as soon as they are found.

Once we do that, we don't need to count them down on exit any more
either.