ps2write - fix PSPageOptions with multi-page input
Commit 7d6d69ff17c43664482fe0dc34676a46ba551d93 accidentally broke
PSPageOptions with multi-page input, because the array of strings was
freed on every page.

This commit fixes that problem; note you cannot change PSPageOptions
once it has been set, except by restoring back to the point before the
device was opened.

Also the documentation formatting was slightly broken, fix that at the
same time.