oss-fuzz 50370: Always fully tidy up a filter pipeline
When shutting down the device, pdfwrite would flush and close the pipeline
of filters in place for the main output stream. But s_close_filters() would
stop the process if a filter in the pipeline threw an error. As pdfwrite
immediately then closes the ouput file, the ultimate target of the pipeline,
this would then later cause a crash then the garbager attempted to "sweep up"
the errant stream objects. Or, in the non-gc use case, it would cause a
memory leak.

So change s_close_filters() to continue to destroy the filters in the pipeline
even after an error (but strill return the relevant error), meaning the filter
pipelines are always cleaned up, even if one or more filters has an error.