Graphics library - enforce limit on Colors with PixelDifference filter
Bug #705776 "stack-overflow in /base/gsmchunk.c:1110:31 in chunk_free_object"

The PixelDifference filter has a memory buffer in the stream state to
hold 'Colors' values, that buffer is hard-coded at 60 values, despite
the fact that, according to the comment in spdiffx.h at line 27 the
value of Colors can be arbitrarily large.

Because the initialisation function doesn't check the value of Colors
we can end up running off the buffer and overwriting memory.

This commit checks the value of Colors and if it exceeds the maximum
value at compile-time returns a rangecheck error.