Color replacement example
Add example of how to perform color replacement
for devn devices for source colors that are
gray, rgb, cmyk, or cielab.  Thanks to Robin
Watts for the suggestion on using the special
ops for this.  This commit adds a new special
op called gxdso_replacecolor.  The PDF14 device
(transparency device) will not allow color
replacement if we are in a soft mask or if
the PDF14 color space is different than the
target device color space.  The psd devices
are set up as a demo to show the use of color
replacement.  See gdevpsd.c and set ENABLE_COLOR_REPLACE
to 1 to activate the example code.  This code
shows the setting of specific device color values
based upon specific source values. Note that access
to the pgs and hence the named color
profile is provided so that a database of colors
mappings can be provided at runtime.

How the memory devices deal with the special_op
procedure was reworked.  Thanks to Robin Watts for
his help on this.  Now any memory special ops should
be handled in mem_spec_op.  If it does not handle it
and a target device exists, then the target device
will have an opportunity to handle the special op.