GhostPDF - fix memory leak with malformed PDF files
Using the file from OSS-fuzz #47753 and this command line:
-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dSAFER -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sDEVICE=nullpage

Leaked memory due to the array checking of the various Box parameters.
One of the MediaBox entries is an (invalid) indirect reference to the
Pages tree, when we fetched the reference from the array we dereferenced
it and stored the dereferenced value in the MediaBox array, thus
creating a circular reference:

Pages->Page->MediaBox
   ^           |
   |------------

Which meant we never counted down any of the objects in the sequence to
0 and freed them.

Using the 'no_store_R' variant avoids storing the dereferenced object
back to the array. Since we don't need the object, other than to get
its numeric value, this doesn't cost any performance.