Coverity 380510/380513: Avoid redundant ID type check
Since we've already checked the font and macro ID types are string, coverity
spots that the type checks in the ID accessor macros are redundant.

So, add accessors for when we already know the ID is a string ID.

Secondly, tidy up macro naming, so it's not so easily confused with a variable
or constant value.