GhostPDF - revise PageLabel handling
This arises from bug #705783 although it does not address the bug
directly.

We should not be processing PageLabels for devices that do not support
writing PageLabel output (basically only pdfwrite). But we had not
picked up the 'WantsPageLabels' parameter from the old PostScript
implementation.

So add that to the device_state structure, and add code to get it from
the device.

After doing that, it transpired that we did not call
pdfi_device_set_flags() until after we had processed the trailer dict.
So add a call nice and early on, in ghostpdf.c, pdfi_init_file().

It then turns out there were several other places where the function
was called, which almost certainly shouldn't be done. It seems likely
that we've moved the call to the function earlier and earlier in the
order of execution as we've discovered we need it, but we really only
(I think) should call it once.

So remove all the other calls.