GhostPDF - detect self-referencing indirect objects
Bug #785703 "stack overflow in pdf/pdf_loop_detect.c:76 pdfi_loop_detector_check_object (exploitable) "

The file is corrupted in such a way that dereferencing object 6 results
in an indirect reference to object 6.

Add a specific check to the dereference code to compare any resulting
indirect reference with the requested object number, and if they are the
same throw a circular reference error.