Coverity ID 380512
We were not using the return value from pdfi_array_get_number(). Since
we're making changes anyway, duplicate the fix in commit
0cbcfd63c0865a8c3a8d9d4bd81ba27cb0432487 to avoid storing the
dereferenced object in the array, in case we end up with a CropBox
where the array contains an indirect reference to the Pages tree.

The earlier commit only updated the array we were using for the canvas
size.