graphics library - prevent seg fault in .tempfile
Coverity ID #380524

A seg fault can be reproduced on Windows using this PostScript:

(d:\\temp) (w+) .tempfile

The problem occurs because on Windows gp_file_name_is_absolute permits
either '\' or '/' to be used as a separator, but gp_file_name_separator
can (obviously) only return one separator, in this case '/'.

This means the loop stripping the prefix fails to find any separator
and exits with plen = -1, which causes a seg fault in the memcpy below.

The first part of the fix is to check plen and if it's less than 0
exit with an error.

The remainder of the commit changes from using the separator string to
calling gp_file_name_check_separator instead. This has two benefits;
firstly it means that on Windows both separators can be detected,
secondly for OS's where the separator is more than a single byte we can
still detect it, which we could not before.

Ghostscript itself never uses .tempfile with a filename, so this can
only occur with PostScript input.