GhostPDF - abort repair of ObjStm on error
OSS-fuzz 50747

The corrupted file has an ObjStm where the value of 'N', the number of
objects in the stream, is corrupted and huge. We read the objects from
a SubFileDecode, so we don't overrun, but we were ignoring the error
return and attempting to read a full 'N' objects.

There's really no point, corrupted ObjStms are more or less irreperable
so if we get an error, just stop.

This doesn't prevent a crash or any other problem, it just means we stop
handling an irreparable file more quickly.