Coverity ID #380567
Turns out there are multiple paths to the error label, many of which
occur before the 'info' enumerator is created. Fix this by initialising
the pointer to NULL and then testing to see if it has been created
before calling end_image. If it hasn't then we didn't call begin_image
(or it failed) and so should not call end_image.