Fix FONTPATH fontmap functionality
Firstly, pdfi is strict in enforcing that dictionary keys must be names, so
when the mapping is a path/file we need to convert that string to a name so we
can store it in the substitute fonts dictionary.

Secondly, an error storing the mapping in the substitute fonts dictionary should
not be propagated - we should continue to use the mapping even if we can't store
it for later reuse.

Stems from (but not fixes) Bug 705831