Bug 705768: Avoid extreme FPs causing out of range ints.
If bbh is large enough, floor(bbh+0.5) can be MIN_INT.

This gives a device height of MIN_INT, which is copied to band_height,
and then trips the band calculations to result in accessing beyond
the end of the band list.

Head all this off, by spotting inst.size.x or y being < 0.