graphics library - bounds check type 2 instruction pointer
This is a speculative fix for OSS-fuzz 50448. Despite the fact that
we cannot reproduce this, it looks like the problem must stem from
reading off the end of the input buffer.

This is possible because we do not check the instruction pointer, but
just keep on incrementing it until we get an endchar.

It looks like we should originally have been doing this, because the
ip_state_t structure, which is used to store the state on a stack
when we call subroutines, has an 'ip_end' pointer. But we were never
using that.....

So add code to set and use it, and add a local variable to hold the
pointer to the end of the input buffer, and check cip against that.