GhostPDF - yet more meddling with /Parent
OSS-fuzz 51011

The annot and optional content code all use the /Parent dictionary, but
doesn't have any circular reference protection.

This duplicates the protection we added to Widget annotations. I intend
to supercede this commit with a more general one which deals in a more
common fashion with access to the /Parent, I just don't want to lose
this commit.