GhostPDF - performance improvement with large CFF CIDFonts
See bug 705534. The interpreter spends an enormous amount of time
checking to see if a dictionary contains a key, or an equally large
amount of time sorting the dictionary after adding a key.

We know that these keys are unique in the dictionary, so this is a waste
of time, lets not do that.