GhostPDF - limit the nesting of arrays and dictionaries
OSS-fuzz 50138

If a file was sufficiently insanely broken to nest arrays or
dictionaries to a sufficient depth, we could exhaust the C execution
stack when trying to free the top level one.

This limits the nesting to a compile-time level (currently 100). If we
exceed that we simply stop nesting the objects but store an error for
reporting on exit.