GhostPDF - check for 0 Size entries in XRefStm /Index array
Bug #705843 " heap-buffer-overflow in pdf/pdf_xref.c:114 read_xref_stream_entries (exploitable)"

The /Index array in the XRefStm holds pairs of values, the first is the
first object number in a subsection, the second is the number of entries
in a subsection.

We were doing loads of checks on the validity of these numbers, but
forgot to check that the size was at least 1. Because we later use
start + size - 1 we could then end up with a ridiculously large index,
because these are unsigned values.

If we get a size of 0, just ignore the subsection.

While making changes, renamed 'end' to 'size' because its a better
descriptive name.