GhostPDF + GS - fix PDFFitPage with square MediaBox
Bug #705849 "New PDF interpreter errors out with PDFFitPage"

If the input PDF file had square media then we would correctly decide
we did not need to rotate the media for a better fit, but we left the
values we would normally use to calculate if rotation was required on
the operand stack, resulting in an error.

Pop the unused values.