GhostPDF - early sanity check on Pages tree root node /Count
A number of OSS-fuzz files timeout due to having an excessively large
/Count (eg 213804087) in the root node of the Pages tree.

This doesn't cause us any real problems, except that it takes a long
time to fail to render that many pages. On the other hand, validating
the entire pages tree for a large file with many nodes could take a
reasonable amount of time. Not huge but it would mean a performance hit
on sensible files just to avoid this penalty on broken files.

As a compromise; this commit checks the root node /Kids array, if it has
sufficient entries to match the /Count then we assume its a flat tree
and the Count is correct. If it has fewer then we assume it's a tree and
we check each of the entries in the /Kids array. If its a leaf node then
we add 1 to the running count. If it is an intermediate node then we add
the /Count of the node to the running count.

If the one-level check matches the root node /Count then we assume the
Count is correct. If it does not, then we assume that the count of
the first level is correct.

We could, if desired, validate the entire tree instead at this point
but I don't think it's worth it. If the tree is really broken then the
file is going to fail. This commit prevents the timeout and if the
corruption is limited to the Count of the root node then it will
recover from that.