Tweak @file quote handling.
The change for bug 703694 in commit 0f38f27a7b3a1eb meant that
we no longer accepted

    -sFoo="Bar Baz"

in an @file, and instead read that as:

   -sFoo="Bar

This seems counterintuitive.

Here we tweak the code so that " is accepted to mean the start
of a quotation after an = sign too.

We also tweak the handling of '\' so that special case handling
only happens in @ files, not from the command line. And we
allow '\"' to mean literal '"'. '\' at the end of the line remains
meaning "continue parsing past the newline". All other cases
of '\' continue to mean literal '\'.