GhostPDF - more Annots and /Parent
OSS-fuzz 51034 (again)

The previous commit fixed the first problem, but I missed the fact that
there was a second problem, masked by the first.

This commit fixes the second problem; the replacement of objects in
arrays/dictionaries can lead to circular references, so we need to
explicitly check the object number of /Parent, not just leave it up to
the dereferencing code.