More type 2 CharString checks for pdfwrite
OSS-fuzz #51275

Two places where the code checks there are sufficient operands on the
stack for the first part of the operation, but does not check for the
second part.

I'm reasonably confident the checks I'm adding are correct, but I'm less
certain about the existing code! It looks like the c2_rlinecurve is
trying to deal with the extra operands. The checks in c2rcurveline don't
look like they even attempt to cater for the final rlineto.