Graphics library - clean up transparency with clist on error
OSS-fuzz #50779 with this command line:

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sDEVICE=pkmraw -r72 -sOutputFile=/temp/out%d.pkm -dBATCH -dNOPAUSE

has a memory leak. This is because the execute_compositor_queue()
function has this comment:

        code = apply_composite(cdev, pgs, mem, pcomp, x0, y0, target); /* Releases the compositor. */

But apply_composite only releases the compositor object on success,
if any of the paths fail it does not free pcomp leading to a memory
leak. Address that here by changing apply_composite so that it does
in fact free pcomp unconditionally.