GhostPDF - fix memory leak with Resources and circular /Parent ref
OSS-fuzz #51034 with this command line:

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sDEVICE=nullpage -r72 -sOutputFile=/temp/out%d.pkm -dBATCH -dNOPAUSE

exhibits large memory leaks. This is because the Parent of one of the
Annots (object 1571) points into the middle of a Fields tree. This
eventuially results in a circular reference.

The code to find resources was using dict_knownget to retrieve the
/Parent, which meant it was being stored in the dictionary which meant
meant we had circular references in the dicts, resulting in the memory
leak.

Do the same as we did for Widget annotations and retrieve Parent dicts
without storing the dereferenced object when searching for Resources.

In addition, we were not counting down the Parent dictionary if we
encountered an error, which meant we still leaked memory. Fix that too.

Finally, a speculative fix, reset code to 0 after we are done finding
Resources in Parents, just in case we end up returning that value.