graphics library - fix memory leak with text enumerator
The file from OSS-fuzz #51198 when run with:

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sDEVICE=nullpage -r72 -sOutputFile=/temp/out%d.pkm -dBATCH -dNOPAUSE

leaks memory because the nullpage device is not released. This turns out
to be for a surprising reason...

When we call gx_default_text_begin that allocates a show enumerator and
initialises it. gx_text_enum_init() increments the reference count of
the device.

Normally we call gs_text_release when we are done with a text
enumerator, which counts down the reference count of the enumerator
and if the reference count reaches 0 calls the rc.free function, in
this case rc_free_text_enum, which calls gx_default_text_release() which
(finally!) decrements the reference count of the device.

Now, if we encounter an error in gx_default_text_begin() after we have
called gs_text_enum_init() then we should free the enumerator before
returning. In many places we don't free it at all, and in the one place
where we do free it, we call gs_free_object() which frees the memory,
but doesn't call any of the reference counting functions and so leaves
the device reference counted up.

At the end of the file we have a reference count of 154 for the device
and so it leaks.

This commit changes the gs_free_object() call into a gs_text_release()
so that we correctly use the reference counting machinery. In addition
a number of other places that could return an error and did not free
the enumerator, now call gs_text_release in order to do so (and count
down the device reference count).