GhostPDF - improve unmatchedmark handling
OSS-fuzz #51318

Commit 78e237d60bf7ba01af6258bc75e250fbcad554e0 to fix OSS-fuzz bug
#51038 caused us to ignore unmatched array or dictionary marks and
simply return, without an error.

That caused a problem with read_dict, which expected either an object
on the stack or an error (not unreasonably) and tried to use a
non-existent object.

This commit amends the previous one. We still ignore unmatched marks,
so we still get the progression noted in the original commit, and the
original OSS-fuzz bug is still fixed, but we now try and find another
token. If we succeed we'll return that, and if we fail we'll return an
error, which satisfies this bug.

Also fix the error message which didn't make grammatical sense.