Further speed improvement in pdfi_dict_put_unchecked()
For very specific circumstances, we know upfront that the keys being written
into a dictionary are unique, hence we added a dictionary "put" function that
doesn't check for duplicates.

In addition, for this application, key/value pairs are never removed from
the dictionary.

So another optimisation is to not search from the first "slot" in the
dictionary to find an empty one, but start from the index of the last entry
we added.

But also include a fallback case, just in case.

This further cuts the interpretation time for fts_23_2309.pdf significantly.