GhostPDF - fix Page count calculation
Commit 1bdc4a87930b1180cebaa73ea8a9cc96c50d34e9 added a check on the
count of the number of pages (to try and avoid timeouts with OSS-fuzz
and huge declared page counts). Unfortunately the wrong dictionary
was queried for the count of intermediate page tree nodes leading to
a wildly incorrect number of pages.

This led to a load of errors being thrown, though all the pages were
rendered correctly.

Correct that here....