Cache bbox device for formatted string annotations
When drawing formatted string annotations, we use the bbox device to work out
the extents of the string we need to draw. Previously this would create and
destroy a bbox device for every such string.

That's quite expensive, especially the creation of the device ICC profiles.

So, cache the device in the context, and reuse it instead.