GhostPDF - performance optimisation for pdfi_stream_to_buffer
The original version of pdfi_stream_to_buffer() always took two passes
over the data; firstly to find the size of the uncompressed data
(PDF files don't normally contain this information) then allocation of
a memory buffer and re-reading the data into the buffer.

Sometimes, however, we know how big we expect the data to be; examples
include Indexed colour lookup tables, threshold arrays and both Mask
and SMask image data.

This commit modifies pdfi_stream_to_buffer() so that if the caller knows
the expected size it can set the 'bufferlen' parameter on call. The
routine will then skip the initial read to determine the size, allocate
the buffer and read the data.

It is possible that the file does not contain enough data to satisy the
request. In that case (return code is ERRC and the initial bufferlen was
not 0) we discard the buffer, reset the expected size to 0, and go back
to the previous method of reading the data twice so that we can return
the actual amount of data available.

This does mean that for files which do not contain as much data as they
should, we will read the data three times, and therefore be slower. But
for conforming files we skip a read, gaining us performance.

The file which inspired this is tests_private/comparefiles/Bug689195.pdf
and that file gets approximately a 15% performance gain from this.