Correct PCL passthrough SEGVs.
In commit 90fa745b7 (Add gpdl support for SmartOffice integration),
I tweaked the way we found the PCL interpreter, so that instead of
relying on it being at a fixed point in the list of interpreters,
we hunted for it by name.

And then I used the wrong name.

Use the right name instead.

I am at a loss to understand why the normal cluster tests didn't
pick this up.