GhostPDF - Fix crash with -dPDFINFO
Stems from bug #705897 (which is not a GS bug report). The file for
that report has a Resource with a Font which points to a non-Font
dictionary.

The font resource checking was incorrectly using a return value of >= 0
from pdfi_dict_knownget_type() to indicate that a Subtype key is
present in the dictionary, the function actually returns > 0 if it is
present and 0 if it is not.

Which caused the code to try and use the non-existent key.