GhostPDF - Some error/warning cases not flagged with PDFSTOPON*
Bug #705900 "Ghostscript - parameter PDFSTOPONWARNING doesn't stop
 on all warnings and errors"

The first case here is actually in the PostScript interpreter and so
does not strictly affect GhostPDF. When executing 'run' on a PDF file
the PostScript code checks to see if the header was at the start of the
file and emits a warning if not. We now check PDFSTOPONWARNING after
that and exit with an error if it is true.

The xref checking (very early part of the new interpreter, predates the
arguments) was repairing files without first checking if it should
exit (if PDFSTOPONERROR is true).