CUPS/PWG/Apple Raster output device: Do not match custom size against PPD
If for a job a custom page size is selected, indicated by something
like "-scupsPageSizeName=Custom.4x8in" or "-scupsPageSizeName=Custom",
do not try to match the requested page size, given by standard
Ghostscript parameters like
"-dDEVICEWIDTHPOINTS=... -dDEVICEHEIGHTPOINTS=...", with the supported
page sizes defined in the PPD file.

This avoids unwished rotations as reported in

    https://github.com/OpenPrinting/cups-filters/issues/484

and also unwished size mismatches if the custom size is very close to
a size defined in the PPD file.