GhostPDF - improve /Pages sanity check
OSS-fuzz #51800

The file has a Pages tree where the root node has an insane number of
pages (> 200 million) *and* the only entry in the Kids array is a
reference to the root node.

This defeats the one level check of the count of pages, because the
Count of the root node matches (unsurprisingly) the count of the
first level, because it points straight back to the root node.

Add a circular reference check and if we trip that just throw an error,
a Pages tree with a circular reference is not going to end well.