Graphics library - fix bug in hex decoding
OSS-fuzz #51789

it is possible to construct a filter chain which includes the ASCII
hex decode filter, and where the filter never terminates. This can
happen if we read the first hex digit of a byte, then ran out of data
from the source trying to read the second byte. This caused the filter
to return 0, which mean we just kept on retrying.

This commit checks to see if we have read one digit (odd_digit is an
input parameter and is assigned to val1, used to detect an existing
digit) and if we have exhausted the input, in which case we return
 1, not 0. This 'seems' to fix the problem but, like all the stream
code, it's hard to be certain this has no unforseen side-effects.

Example PostScript code to reproduce:

%!PS
/TFile currentfile <</EODCount 0 /EODString (endstream) >> /SubFileDecode filter /ASCIIHexDecode filter /DCTDecode filter def
{TFile read {==}{exit} ifelse} loop
ffd8ffe000104a46494600010100000100010000ffdb004300030202020202030202020303030304
060404040404080606050609080a0a090809090a0c0f0c0a0b0e0b09090d110d0e0f101011100a0c
12131210130f101010ffdb00430103030304030408040408100b090b101010101010101010101010
1c0
0011080001000103011100021101031101ffc40014000100000000000000000000000000000008ff
c40014100100000000000000000000000000000000ffc40015010101000000000000000000000000
00000709ffc40014110100000000000000000000000000000000ffda000c03010002110311003f00
3a03154dffd9
endstream

(Exit) == flush